requirements/prod.txt
flask 
flask-sqlalchemy 
flask-bcrypt
python-dotenv
flask-session
redis
requests
beautifulsoup4